Relocate weighted primitives accordingly
They should no longer be in the jung package
